+++ lib/bloc/timer_graph/timer_graph_state.dart
